// program/clients/typescript/src/generated/types/chargebackPolicy.ts
/**
 * This code was AUTOGENERATED using the codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import { combineCodec, getStructDecoder, getStructEncoder, getU64Decoder, getU64Encoder, type FixedSizeCodec, type FixedSizeDecoder, type FixedSizeEncoder } from '@solana/kit';



export type ChargebackPolicy = { maxAmount: bigint; maxTimeAfterPurchase: bigint;  };

export type ChargebackPolicyArgs = { maxAmount: number | bigint; maxTimeAfterPurchase: number | bigint;  };




export function getChargebackPolicyEncoder(): FixedSizeEncoder<ChargebackPolicyArgs> {
  return getStructEncoder([['maxAmount', getU64Encoder()], ['maxTimeAfterPurchase', getU64Encoder()]]);
}



export function getChargebackPolicyDecoder(): FixedSizeDecoder<ChargebackPolicy> {
  return getStructDecoder([['maxAmount', getU64Decoder()], ['maxTimeAfterPurchase', getU64Decoder()]]);
}




export function getChargebackPolicyCodec(): FixedSizeCodec<ChargebackPolicyArgs, ChargebackPolicy> {
  return combineCodec(getChargebackPolicyEncoder(), getChargebackPolicyDecoder());
}