// program/clients/typescript/src/generated/types/settlementPolicy.ts
/**
 * This code was AUTOGENERATED using the codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getBooleanDecoder,
  getBooleanEncoder,
  getStructDecoder,
  getStructEncoder,
  getU32Decoder,
  getU32Encoder,
  getU64Decoder,
  getU64Encoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from '@solana/kit';

export type SettlementPolicy = {
  minSettlementAmount: bigint;
  settlementFrequencyHours: number;
  autoSettle: boolean;
};

export type SettlementPolicyArgs = {
  minSettlementAmount: number | bigint;
  settlementFrequencyHours: number;
  autoSettle: boolean;
};

export function getSettlementPolicyEncoder(): FixedSizeEncoder<SettlementPolicyArgs> {
  return getStructEncoder([
    ['minSettlementAmount', getU64Encoder()],
    ['settlementFrequencyHours', getU32Encoder()],
    ['autoSettle', getBooleanEncoder()],
  ]);
}

export function getSettlementPolicyDecoder(): FixedSizeDecoder<SettlementPolicy> {
  return getStructDecoder([
    ['minSettlementAmount', getU64Decoder()],
    ['settlementFrequencyHours', getU32Decoder()],
    ['autoSettle', getBooleanDecoder()],
  ]);
}

export function getSettlementPolicyCodec(): FixedSizeCodec<
  SettlementPolicyArgs,
  SettlementPolicy
> {
  return combineCodec(
    getSettlementPolicyEncoder(),
    getSettlementPolicyDecoder()
  );
}